[#43] create internal library for types
Resolves #43
