+++ elm-street.cabal
@@ -60,14 +60,48 @@ library
                        TypeApplications
                        ViewPatterns
 
-executable generate-elm
-  hs-source-dirs:      generate-elm
-  main-is:             Main.hs
+library types
+  hs-source-dirs:      types
+  exposed-modules:     Types
 
   build-depends:       base >= 4.11 && < 4.13
                      , elm-street
                      , text
                      , time
+
+  ghc-options:         -Wall
+                       -threaded
+                       -rtsopts
+                       -with-rtsopts=-N
+                       -Wincomplete-uni-patterns
+                       -Wincomplete-record-updates
+                       -Wcompat
+                       -Widentities
+                       -Wredundant-constraints
+                       -fhide-source-paths
+
+  default-language:    Haskell2010
+  default-extensions:  ConstraintKinds
+                       DeriveGeneric
+                       GeneralizedNewtypeDeriving
+                       InstanceSigs
+                       KindSignatures
+                       LambdaCase
+                       OverloadedStrings
+                       RecordWildCards
+                       ScopedTypeVariables
+                       StandaloneDeriving
+                       TupleSections
+                       TypeApplications
+                       ViewPatterns
+
+executable generate-elm
+  hs-source-dirs:      generate-elm
+  main-is:             Main.hs
+
+  build-depends:       base
+                     , elm-street
+                     , types
 
   ghc-options:         -Wall
                        -threaded
